api/requirements.txt
Flask==2.3.3
mysql-connector-python==8.1.0
pymongo==4.5.0
pika==1.3.2